Implement mt_srand_array
The new function mt_srand_array is used to seed the MT19937 PRNG
with more than 32 bits. The function takes an array of integers
and seeds the PRNG. The seeding procedure is compatible with
init_by_array() in mt19937ar.c by MT19937 authors.
If an input value is not a 32-bit integer (long or double),
a warning will be generated and the PRNG will not be seeded.