Implement mt_srand_array
The new function mt_srand_array is used to seed the MT19937 PRNG
with more than 32 bits. The function takes an array of integers
and seeds the PRNG. The seeding procedure is compatible with
init_by_array() in mt19937ar.c by MT19937 authors.
All input values that are not 32-bit integers shall be silently
converted and/or truncated to 32-bit integers.